+++ http/write_handler.go
@@ -125,7 +125,7 @@ func (h *WriteHandler) handleWrite(w http.ResponseWriter, r *http.Request) {
 		bucket = b
 	}
 
-	if !platform.Allowed(platform.WriteBucketPermission(bucket.ID), auth) {
+	if !auth.Allowed(platform.WriteBucketPermission(bucket.ID)) {
 		EncodeError(ctx, errors.Forbiddenf("insufficient permissions for write"), w)
 		return
 	}
